Reintroduce the BigMod function for computing modulo operations
The decNumber library requires a context with enough digits to cover the full
range of the two values being used in the modulo operation.  Specifically,
operations like 1e1000 mod 1e-1000 require a context with a bit over
2000 digits to get the correct result.

The big-mod function was lost along the line and this reinstates it.
We might want to extend the number of digits to cover our entire range of
reals (i.e. 20000+some digits).

Also convert integers to real1071's to avoid losing any digits.

Closes #213